WinGui: Update the libhb scan code to work with the updated API. See rev6602 for API changes.
git-svn-id: svn://svn.handbrake.fr/HandBrake/trunk@6608 b64f7644-9d1e-0410-96f1-a4d463321fa5

+++ win/CS/HandBrake.ApplicationServices/Services/Scan/LibScan.cs
@@ -476,7 +476,7 @@ namespace HandBrake.ApplicationServices.Services.Scan
                         TitleNumber = title.TitleNumber,
                         Duration = title.Duration,
                         Resolution = new Size(title.Resolution.Width, title.Resolution.Height),
-                        AspectRatio = title.AspectRatio,
+                        AspectRatio = (decimal)title.AspectRatio,
                         AngleCount = title.AngleCount,
                         ParVal = new Size(title.ParVal.Width, title.ParVal.Height),
                         AutoCropDimensions = title.AutoCropDimensions,
