+++ win/CS/HandBrake.Interop/HandBrakeInterop/SourceData/Title.cs
@@ -101,7 +101,7 @@ namespace HandBrake.Interop.SourceData
 		/// <summary>
 		/// Gets or sets the aspect ratio of this Title
 		/// </summary>
-		public double AspectRatio { get; set; }
+		public decimal AspectRatio { get; set; }
 
 		/// <summary>
 		/// Gets or sets the number of angles on the title.
